move watcher from repo to domain
